spy_blog/requirements.txt
alembic==1.4.2
aniso8601==8.0.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-Limiter==1.4
Flask-Mail==0.9.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gevent==20.6.2
greenlet==0.4.16
gunicorn==19.8.1
hashids==1.3.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
limits==1.5.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psutil==5.7.2
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
urllib3==1.25.10
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
